requirements.txt
git+https://github.com/a5c-ai/mcpml#egg=mcpml
playwright
PyYAML